pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dlib>=20.0.0",
    "face-recognition>=1.3.0",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
]
